Fixed clipboard site login issue
Due to HTML updates to the clipboard site that was previously being used, the bot was no longer able to login using the same element ids, so an update to the login instruction was required. After some work, it seemed like that website no longer met the bot's requirements, so a new website needed to be used, with all new instructions.